+++ t-sparqlUpdate/pom.xml
@@ -16,7 +16,7 @@
 	<packaging>bundle</packaging>
 
 	<properties>
-		<dpu.package>cz.cuni.mff.xrg.uv.transformer.sparql.update</dpu.package>
+		<dpu.package>eu.unifiedviews.plugins.transformer.sparql.update</dpu.package>
 		<dpu.mainClass>SparqlUpdate</dpu.mainClass>
 	</properties>
